アラート部分リファクタリング
