+++ src/Card.tsx
@@ -7,6 +7,8 @@ import DeleteIcon from '@material-ui/icons/Delete';
 import CheckBoxOutlinedIcon from '@material-ui/icons/CheckBoxOutlined';
 import CheckBoxOutlineBlankOutlinedIcon from '@material-ui/icons/CheckBoxOutlineBlankOutlined';
 import {CardsContext} from './Container'
+
+
 const style = {
   border: '1px dashed gray',
   padding: '0.5rem 1rem',
