+++ src/Container.tsx
@@ -1,31 +1,15 @@
-import React,{ FC, useState, useCallback,useEffect} from 'react'
+import React,{ FC, useState, useCallback} from 'react'
 import { Card } from './Card'
 import update from 'immutability-helper'
-
 import AddIcon from '@material-ui/icons/Add';
-import CheckBoxOutlinedIcon from '@material-ui/icons/CheckBoxOutlined';
 import CheckBoxOutlineBlankOutlinedIcon from '@material-ui/icons/CheckBoxOutlineBlankOutlined';
-import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
-import Alert from '@material-ui/lab/Alert';
-import IconButton from '@material-ui/core/IconButton';
-import CloseIcon from '@material-ui/icons/Close';
-import Collapse from '@material-ui/core/Collapse';
-// アラート
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      width: '100%',
-      '& > * + *': {
-        marginTop: theme.spacing(2),
-      },
-    },
-  }),
-);
+import  AlertComponent from './components/alert'
+
+
 export interface Item {
   id: number
   text: string
 }
-
 export interface ContainerState {
   cards: Item[]
 }
@@ -35,7 +19,6 @@ export const Container: FC = () => {
     const [task, setTask] = useState("");
     const [cards, setCards] = useState([{ id: 0, name: "" },]);
     const [message, setMessage] = useState("")
-    const classes = useStyles();
     const [alert, setAlert] = useState("");
     // インプットタグ入力フォーム値取得
     const handleChange = (event:React.ChangeEvent<HTMLInputElement>) => {
@@ -55,7 +38,6 @@ export const Container: FC = () => {
         setAlert("error")
         setMessage("空欄のタスクは登録できません！")
       }
-     
     }
 
     // エンタキー押した時にタスク追加
@@ -96,7 +78,10 @@ export const Container: FC = () => {
       )
     }
     }
-
+    // アラート初期化メソッド
+    const AlertResetFunc=()=>{
+      setAlert("")
+    }
     return (
   <>
   <div className= "Container">
@@ -105,49 +90,13 @@ export const Container: FC = () => {
     <div className="headerItem" onClick={()=> DeleteAllTask()}>全削除</div>
   </div>
   <hr/>
-  <div className={classes.root}>
-{
-alert=="error"?  <Alert severity="error"
-action={
-  <IconButton
-    aria-label="close"
-    color="inherit"
-    size="small"
-    onClick={() => {
-      setAlert("");
-    }}
-  >
-    <CloseIcon fontSize="inherit" />
-  </IconButton>
-}>
-  {message}
-</Alert>:null
-}
-{
-alert=="success"?  <Alert severity="success"
-action={
-  <IconButton
-    aria-label="close"
-    color="inherit"
-    size="small"
-    onClick={() => {
-      setAlert("");
-    }}
-  >
-    <CloseIcon fontSize="inherit" />
-  </IconButton>
-}>
-  {message}
-</Alert>:null
-}
-    </div>
-   <CardsContext.Provider value={{cards,setCards}}>
+    <AlertComponent message={message} alert={alert}  AlertResetFunc={()=>AlertResetFunc()}/>
+    <CardsContext.Provider value={{cards,setCards}}>
     {cards.length!=1?
     <div>{cards.map((card:any , i:number) => renderCard(card, i))} </div>
     :null} 
     {/* inputタグ */}
     <div className="Input">
-
       < CheckBoxOutlineBlankOutlinedIcon />
         <div className="cp_iptxt">
         <label className="ef">
@@ -160,7 +109,7 @@ action={
       <AddIcon className="addIcon" onClick={()=>addtask()}/>
     </div>
     </CardsContext.Provider>
-    </div>
+</div>
   </>
     )
   }
